Configuration done with cached data.
The data provided by Lavoisier is a bit too complex to handle neatly
directly inside Ansible. The structure is too deep to be able to
efficiently loop over, and the json query becomes difficult to read.
For these reasons, I've decided to parse the data out of band with a
small python script in files/ which writes a cached data file in
JSON format.