Add the possiblity to restrict the json query when creating the VOMS files (#26)
# Summary

When using this module on a site, it takes a long time to go through all
the VOs. This pull request permits to restrict the VOMS file generation
to the VOs supported by the site. For doing this, the site admin should
only set the `supported_vos` variable. ie.:
```
supported_vo:
  - biomed
  - dteam
  - ops
```
If the `supported_vos` is not set, then all the VOMS files are created
(the default behaviour of this module is not modified).

Fix #25

Co-authored-by: Enol Fernández <enol.fernandez@egi.eu>
Co-authored-by: Baptiste Grenier <baptiste.grenier@egi.eu>